Cetacean Dataset

A dataset containing whales and dolphins that spent some period of time captive in the US between 1938 and May 7, 2017. See: https://pudding.cool/2017/07/cetaceans/

Usage

cetacean_data

Format

A tibble with 2194 rows and 20 columns:

species

The common name for an individual's species

id

Identification numbers assigned to an animal

name

An individual animal's name

sex

An individual animal's sex

accuracy

The accuracy of an animal's birth date, a = actual date, e = estimated date, u = unknown

birthYear

The year that an individual animal was born

acquisition

The method through which an animal was acquired

originDate

The year that an animal entered captivity

originLocation

The location that an animal originated from

mother

The name of an animal's biological mother

father

The name of an animal's biological father

transfers

A list of facilities that an animal was transferred to, with the approximate dates that they were transferred

currently

The location where a living animal currently resides or last known location

region

The region of the world where the animal either currently or most-recently lived

status

The current status of an animal.

statusDate

The date that an animal's status changed

COD

The animal's reported cause of death.

notes

Additional notes e.g names

transferDate

The date an animal was transferred into the US

transfer

The types of transfers that an animal has been involved with

entryDate

The date that an individual animal entered the US captive population.

Source

https://github.com/rfordatascience/tidytuesday/tree/master/data/2018/2018-12-18